Update and improve project documentation
Also merge file-upload and news-upload api endpoints into a single
upload module instead of being independent.
