+++ api/application.py
@@ -15,7 +15,7 @@ logging.getLogger("haystack").setLevel(logging.INFO)
 
 
 def get_application() -> FastAPI:
-    application = FastAPI(title="Haystack-API", debug=True, version="0.1", root_path=ROOT_PATH)
+    application = FastAPI(title="Mapintel-API", debug=True, version="0.1", root_path=ROOT_PATH)
 
     # This middleware enables allow all cross-domain requests to the API from a browser. For production
     # deployments, it could be made more restrictive.
