+++ ui/README.md
@@ -1,44 +1,25 @@
-TODO: ADAPT TO THIS PROJECT
-## Demo UI
+# Mapintel UI
 
-This is a minimal UI that can spin up to test Haystack for your prototypes. It's based on streamlit and is very easy to extend for your purposes. 
+This is a streamlit-based UI application for interactive semantic searching and exploration of a large collection of news articles.
 
-![Screenshot](https://raw.githubusercontent.com/deepset-ai/haystack/master/docs/_src/img/streamlit_ui_screenshot.png)
+![Screenshot](../outputs/figures/ui_screenshot.png)
 
 ## Usage
 
-### Get started with Haystack
+### Option 1: Container
 
-The UI interacts with the Haystack REST API. To get started with Haystack please visit the [README](https://github.com/deepset-ai/haystack/tree/master#key-components) or checko out our [tutorials](https://haystack.deepset.ai/docs/latest/tutorial1md).
+Just run
+```
+docker-compose --profile ui up
+``` 
+in the root folder of the Mapintel repository. This will start three containers (Open Distro for Elasticsearch, FastAPI application, UI application).
+You can find the UI application at `https://localhost:8501`
 
-### Option 1: Local
+### Option 2: Local
 
 Execute in this folder:
 ```
 streamlit run webapp.py
 ```
 
-Requirements: This expects a running Haystack REST API at `http://localhost:8000`
-
-### Option 2: Container
-
-Just run
-```
-docker-compose up -d
-``` 
-in the root folder of the Haystack repository. This will start three containers (Elasticsearch, Haystack API, Haystack UI).
-You can find the UI at `http://localhost:8501`
-
-## Evaluation Mode
-
-The evaluation mode leverages the feedback REST API endpoint of haystack. The user has the options "Wrong answer", "Wrong answer and wrong passage" and "Wrong answer and wrong passage" to give feedback. 
-
-In order to use the UI in evaluation mode, you need an ElasticSearch instance with pre-indexed files and the Haystack REST API. You can set the environment up via docker images. For ElasticSearch, you can check out our [documentation](https://haystack.deepset.ai/docs/latest/documentstoremd#Initialisation) and for setting up the REST API this [link](https://github.com/deepset-ai/haystack/blob/master/README.md#7-rest-api).
-
-To enter the evaluation mode, select the checkbox "Evaluation mode" in the sidebar. The UI will load the predefined questions from the file [`eval_lables_examles`](https://raw.githubusercontent.com/deepset-ai/haystack/master/ui/eval_labels_example.csv). The file needs to be prefilled with your data. This way, the user will get a random question from the set and can give his feedback with the buttons below the questions. To load a new question, click the button "Get random question". 
-
-The file just needs to have two columns separated by semicolon. You can add more columns but the UI will ignore them. Every line represents a questions answer pair. The columns with the questions needs to be named “Question Text” and the answer column “Answer” so that they can be loaded correctly. Currently, the easiest way to create the file is manully by adding question answer pairs. 
-
-The feedback can be exported with the API endpoint `export-doc-qa-feedback`. To learn more about finetuning a model with user feedback, please check out our [docs](https://haystack.deepset.ai/docs/latest/domain_adaptationmd#User-Feedback).
-
-![Screenshot](https://raw.githubusercontent.com/deepset-ai/haystack/master/docs/_src/img/streamlit_ui_screenshot_eval_mode.png)
+**Requirements**: This expects a running Fast API application at `https://localhost:8000` and a running Open Distro for Elasticsearch instance at `https://localhost:9200`. Also, all python and system dependencies must be satisfied.
